Replace all current with next modules
