+++ p2panda-rs/src/test_utils/fixtures/mod.rs
@@ -7,7 +7,6 @@
 //!
 //! Implemented using the [`rstest`](https://github.com/la10736/rstest) library.
 mod db_fixtures;
-pub mod db_fixtures_next;
 mod document_fixtures;
 mod entry_fixtures;
 mod hash_fixtures;
